requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.5
fastapi==0.135.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.191.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
joblib==1.5.3
llvmlite==0.46.0
numba==0.64.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
PyJWT==2.11.0
PyNaCl==1.6.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
